Adding surety system (DelMap mode)

+++ src/game/handler/del_map.rs
@@ -1,38 +1,52 @@
-use crate::error::{GameError, GameResult};
+use crate::error::GameResult;
 use crate::game::handler::utils::disp_list;
 use crate::utils::maps_list;
 use crate::MAPS_PATH;
 
-use std::fs::remove_file;
-use std::io::{stdin, stdout, Write};
-use std::path::Path;
+use std::fs::{remove_file, DirEntry};
+use std::io::{stdin, stdout, Stdin, Stdout, Write};
+use std::path::{Path, PathBuf};
 
-fn ask_idx() -> GameResult<usize> {
-    let mut stdout = stdout();
-    let stdin = stdin();
-
+fn ask(maps: &[DirEntry], stdin: &Stdin, stdout: &mut Stdout) -> GameResult<PathBuf> {
     loop {
         print!("Choose the map to remove by giving an index: ");
         stdout.flush()?;
         let mut out = String::new();
         stdin.read_line(&mut out)?;
 
-        match out.trim().parse::<usize>() {
-            Ok(idx) => break Ok(idx - 1),
+        let idx = match out.trim().parse::<usize>() {
+            Ok(idx) => idx - 1,
             Err(_) => continue,
         };
+
+        match maps.get(idx) {
+            Some(entry) => break Ok(entry.path()),
+            None => continue,
+        }
     }
 }
 
+fn is_sure(stdin: &Stdin, stdout: &mut Stdout) -> GameResult<bool> {
+    print!("Are you sure? (\"yes\"/\"no\"): ");
+    stdout.flush()?;
+    let mut out = String::new();
+    stdin.read_line(&mut out)?;
+    Ok(out.trim() == "yes")
+}
+
 pub fn run() -> GameResult<()> {
+    let stdin = stdin();
+    let mut stdout = stdout();
+
     let maps = maps_list(Path::new(MAPS_PATH))?;
 
     disp_list()?;
     println!("--- Deleting a map ---\n");
-    let idx = ask_idx()?;
+    let map_path = ask(&maps, &stdin, &mut stdout)?;
 
-    let map_path = maps.get(idx).ok_or(GameError::IncorrectInput)?.path();
-    remove_file(map_path)?;
+    if is_sure(&stdin, &mut stdout)? {
+        remove_file(map_path)?;
+    }
 
     Ok(())
 }
